renamed User component to UserPage
