+++ webapp/src/pages/index.ts
@@ -1,6 +1,6 @@
 export { default as Landing } from './Landing/Landing';
 export { default as Users } from './Users/Users';
-export { default as User } from './User/User';
+export { default as UserPage } from './UserPage/UserPage';
 export { default as Play } from './Play/Play';
 export { default as Chat } from './Chat/Chat';
 export { default as ComponentsBook } from './ComponentsBook/ComponentsBook';
